Removed the Spring DM classes bundles since the namespaces are leaking from the Spring 2.5.5 area into Spring 2.0.6
At the moment I am not certain if this is an issue is not. It appears that there isnt isolation on classloader.getResources for things like
getResources("/META-INF/x/y/z/someconfig.properties") This appears to go over all bundles regardless of version bindings. Perhapse becuase there is no
version information attached to resource files outside the package space.


git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@53690 fdecad78-55fc-0310-b1b2-d7d25cf747c9