This versin almost starts the kernel up.
It contains a patch to the Component Manager that solves some of the resolution issues with resources between classloaders and bean resolution.
This is not complete as I am still getting bean resolution issues.

It looks like use of the thread context classloader is causing problems as that is not the classloader of the bundle, there is still work to do on this.


git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@53696 fdecad78-55fc-0310-b1b2-d7d25cf747c9