Reviewed the way ThreadLocalManager works and merged into a single Cache Manager with scope.
There are 4 types of cache the cache manager can create. Request scoped that last for the request cycle,
thread scoped that bind to the thread, instance scoped that are shared between thread on the same JVM, and
cluster scoped (with or without replication) that invalidate and bind over the cluster. ThreadLocalManager provided a
request scoped cache.

In the current implementation, request scoped and thread scoped dont invalidate, but request scoped caches are very short lived.
I may remove thread scoped as there appears to be little need for it and it feels dangerous.

When using these caches the caller requests a cache for their purposes, the key namespace inside the cache is private to the users
of that cache. IMHO this will connect well to transation and various other things. Items in the thread bound caches can still watch
for eviction from the cache.


git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@54197 fdecad78-55fc-0310-b1b2-d7d25cf747c9