Added Integration Test, kernel and JCR now starting inside full lifecycle.
Components are located by building a classloader and then locating all the SAKAI-INF/component.xml files in the classpath.
This list is converted into a set of ComponentSpecs that this then loaded by the component manager.

This task is performed by the ComponentLoaderService, which is invoked in the kernel bootstrap phase.

I now need a good way of building that list of jars to load, as in a normal environment this will not be in a classloader visible to
the kernel class loader. I am not certain what happens to classloaders when they are thrown away, and perhapse jar scanning might be a better way of doing this.


git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@54240 fdecad78-55fc-0310-b1b2-d7d25cf747c9