Fixed some classloader issues for shared and component classloaders added more unit tests to get fuller coverage inside the kernel
Still need to think about how we are specifing components in the component spec. It may make much more sense to use
the same model as maven since components are not just jars with a component spec.

Also, the Maven2 dependency resover currently doesnt pull dependencies from remote repositories, it might make sense to make this
possible, then all that would be required to start up the kernel will be the component jars.


git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@54348 fdecad78-55fc-0310-b1b2-d7d25cf747c9