On reflection the caching UserEnvrionment based on sesssion ID makes no sense becuase we are forced to invalidate the cache based on userID, and so any state that is bound to session ID rather than user ID will be lost, eliminating the value of aquiring that state in the first place.
In short, since we update User Environment, we cant aquire transient state on a session by session basis.

This patch removes the complexities introduced in the SessionManager surrounding tracking sessions to users,
eliminates the need to have a session for every request, and changes the binding of user environment to bind
to user id rather than sesssion.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@56306 fdecad78-55fc-0310-b1b2-d7d25cf747c9