Added in the ability to have an externally provided SubjectStatement or ACS.
This means that if an ACS exists on an object with a type or SubjectType.PROVIDED the Registry SubjectStatement.REGISTRY_PROVIDER will be
used to get a list of SubjectTokenProvider(s) which will be consulted in turn for a match. The first match will
result in a match for the SubjectStatement (or ACS) and the appropriate action will be taken.

This code is *only* invoked if the target object contains a ACS with SbjectType == PROVIDED, otherwise its not invoked.

It would be possible to write a rules engine as an external provided type.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@56570 fdecad78-55fc-0310-b1b2-d7d25cf747c9