KERN-49
Patch From Aaron Zeckoski

Here is a patch which will add 2 -P options for running:
run-only - just runs the server as is
overlay-run-only - overlays the kernel code and then runs

The default behavior of run is not changed

There is also a change to the start.sh file which will allow someone to set the SAKAI_KERNEL_PROPERTIES file. As it is currently, it cannot be set because the start.sh will override any setting that is made. This will now check to see if t it is set and then set it if not.

Thanks

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57102 fdecad78-55fc-0310-b1b2-d7d25cf747c9