KERN-48
Fixed,
There is a property core.component.locations that is defaulted to the kernel.
This can be overridden in local components and if you set
component.locations they are appended to the end of teh core components

THere is also annother feature that has always been there, any property starting with + is
appended to that property rather than over writing it.

This patch causes some problems for unit tests that were orriginally managing the kernel explicitly.
You MUST clear core.component.locations if you are expecting the kernel to come from the classpath otherwise
the kernel will start up twice. Once from the configured core.component.locations and once from the classpath.

The KernelIntegrationBase.beforeClass() and KernelIntegetionBase.afterClass() perform these functions corrrectly, or you
can DIY if you want with KernelIntegrationBase.disableKernelStartup() and KernelIntegrationBase.enableKernelStartup()
methods.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57106 fdecad78-55fc-0310-b1b2-d7d25cf747c9