KERN-72
Fixed,
The JPA is checked for the existance of the site correctly, and the transactions are managed correctly.

When converting a byte array into a string you cannot assume it will be UTF-8 you must specify it.
Also, if an input stream is opened, it must be closed, even if its a byte array, dont wait for the finalizer to do it.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57355 fdecad78-55fc-0310-b1b2-d7d25cf747c9