KERN-64
First part of this is done. There is a SecureSakaiAccessManager.
The creation code for the session had to be modified since it needed the Guice injector.
The settings in the RepositoryConfig.xml file are ignored for the AccessManager and Guice now provides the impl of this interface so config is in KernelModule. Might do the same for LoginManager.

The SecureSakaiAccessManager is not active, since more testing is required so at the moment the only active change is the way in which the AccessManage is instanced.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57356 fdecad78-55fc-0310-b1b2-d7d25cf747c9