KERN-70
Discovered dome problems with the EntityManager which must not be shared between threads. The transaction is a singleton on the entity manager so
2 threads with the same entity manager will commit the transaction randomly.

The solution is to bind the entity manager to a request scoped cache (keeping track of all the EM's in use) in a ReferenceMap.

The EM's get created lazilly with a < 1ms overhead.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57422 fdecad78-55fc-0310-b1b2-d7d25cf747c9