KERN-88
The indexing mechanism is complete, In the RestPatchProvider there is a parameter 'i' which if specified is an
array of 0 and 1 (multiple POST params) that defines if the corresponging key is indexed or not.
Once Indexed, they param can only be removed from the index by removing the param and re-adding.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57732 fdecad78-55fc-0310-b1b2-d7d25cf747c9