KERN-98
This fixes sorting in search, however there are some issues.
The properties node have to be attached to the jcr:content node and not the nt:file node.
jcr:content is now nt:unstructured rather than nt:resource (same as webav) to enable
adding the properties nodes at that possition.
The output on search merges properties from nt:file and jcr:content nodes
when I say nt:file and jcr:content I am mixing terms. the jcr:content is the name of the child node of
a nt:file node.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57894 fdecad78-55fc-0310-b1b2-d7d25cf747c9