KERN-95
Added Friends service with storage on JCR.
At the moment, there are a number of things missing that matter.
There is no sorting or paging of the friends lists, partially because we dont store the uder information in
the list. I do need to think of the most efficient way of sortign this list because its stored inside a single json file.
We probably need to re-sort when we save the list, this will make edits slow, but avoid re-sorting on every request.
It will also make it slow for changes to the sort order due to edits on firstname and last name to come through. How often
to users change their names ?

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@57935 fdecad78-55fc-0310-b1b2-d7d25cf747c9