KERN-108
Fixed error on first use of webdav, by making the logout logout correctly from the session on the thread.
Also discovered that webdav is not wrapped in a sakai servlet
This is still there, but we may need to prevent sessions from being created now that the filter is there.

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@58089 fdecad78-55fc-0310-b1b2-d7d25cf747c9