KERN-197
Patch from  	 David Haines

The bootstrap pom should support setting arguments to pass to the tomcat instance that it runs in order to allow:
- running the tomcat in jdpa mode for debugging
- sending output to catalina logs.

The attached patch adds an argument that can be used on the maven command line to specify the arguments to tomcat. The value defaults to "run". To run with a remote debugger add -Dtomcat.args="jpda start". This will also send the logging to the catalina log files.

Thanks

git-svn-id: https://source.sakaiproject.org/contrib/k2/trunk@58627 fdecad78-55fc-0310-b1b2-d7d25cf747c9